Consolidate pass updater and entity utils
